+++ changelog.md
@@ -1,3 +1,7 @@
+# v0.0.2
+- Setup register endpoint
+- Added Joi module for json validation
+
 # v0.0.1
 Initial Commit
 - Setup app index, packages, default route